Move the calendar event call recording tab upgrade command to 2-31 after the 2.30 release
